fixing the cost button
I had an error where my varaibles weren't displaying correctly or just wrong.
I fixed this in a few ways, where I added a new global varaible that I could keep track of easier, and then double checking that that variable was defined correctly in my code.
I also added a rounding function to the cost, as the way I implimented it caused it to slowly add more and more decimal places but round() fixed that